pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tonic-fabricate"
version = "1.0.1"
description = "The official Fabricate client for Python"
authors = [{name = "Fabricate Tools"}]
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",

    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.25.0",
    "tomli>=1.2.0; python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "python-dotenv>=0.19.0",
    "pytest>=6.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.900",
]

[project.urls]
Homepage = "https://github.com/fabricate-tools/client-python"
Repository = "https://github.com/fabricate-tools/client-python"
Documentation = "https://github.com/fabricate-tools/client-python#readme"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 